+++ librelane/scripts/openroad/common/io.tcl
@@ -119,7 +119,7 @@ proc read_current_sdc {} {
     }
 
     if { ![string_in_file $::env(_SDC_IN) "set_propagated_clock"] && ![string_in_file $::env(_SDC_IN) "unset_propagated_clock"] } {
-        if { [info exists ::env(OPENLANE_SDC_IDEAL_CLOCKS)] && $::env(OPENLANE_SDC_IDEAL_CLOCKS) } {
+        if { [info exists ::env(LIBRELANE_SDC_IDEAL_CLOCKS)] && $::env(LIBRELANE_SDC_IDEAL_CLOCKS) } {
             puts "\[INFO\] No information on clock propagation in input SDC file-- unpropagating all clocks."
             unset_propagated_clock [all_clocks]
         } else {